Implements prune functionality
Adds the ability to remove unmanaged resources (volumes, networks, containers)
that are labeled with the identifier but not present in the current
configuration.

Adds a `--prune` flag to both `plan` and `apply` commands. The `plan`
command now displays resources to be removed when the flag is used.
When prune is not set but resources will be removed, a message is displayed.